indexer/requirements.txt
kafka-python==2.0.2
beanie==1.10.8
motor==2.5.1
pydantic==1.9.0
